+++ packages/bezier-codemod/src/transforms/v2-z-index-interpolation-to-css-variable/fixtures/z-index-enum-not-from-bezier-react.input.tsx
@@ -1,0 +1,5 @@
+import { ZIndex } from 'some-library'
+
+export const OVERLAY_POSITION1 = {
+  zIndex: ZIndex.Modal,
+}
